+++ CHANGELOG.rst
@@ -4,7 +4,11 @@ Changelog
 
 Unreleased
 ==========
-* feat: Add AliasContent admin Site filter
+* feat: Add AliasContent admin changelist Site filter
+
+1.2.0 (2022-04-11)
+==================
+* feat: Changed Site dropdown url for Aliases to point to the admin changelist
 * feat: Refactor alias topdown and versioning action menus
 
 1.1.0 (2022-04-06)